requirements.txt
flask
requests
websockets
pillow
numpy
face_recognition
opencv-python
azure-storage-blob